Improve Widget functionality and interface aesthetics
The Widget functionality has been changed to show a room image upon selection, which is hidden initially, and the function 'initializeSeafloor' now replaces 'initializeCalendar'. The image also fades in for a better visual effect, which is handled inside a callback. Changes have been made in HTML file to adjust layout, and in CSS file to style the submit button.